refactor: delete extra files from app
