+++ apps/designer/components/LoadingBar.tsx
